moving check scripts
