+++ applications_superstaq/__init__.py
@@ -6,6 +6,7 @@ from applications_superstaq.superstaq_exceptions import (
     SuperstaQNotFoundException,
     SuperstaQUnsuccessfulJobException,
 )
+from . import check
 from . import converters
 from . import finance
 from . import logistics
@@ -22,6 +23,7 @@ __all__ = [
     "SuperstaQModuleNotFoundException",
     "SuperstaQNotFoundException",
     "SuperstaQUnsuccessfulJobException",
+    "check",
     "converters",
     "finance",
     "logistics",
